+++ services/learn-card-network/learn-cloud-service/src/app.ts
@@ -1,13 +1,19 @@
 import { t } from '@routes';
-import { customStorageRouter } from '@routes/customStorage';
-import { indexRouter } from '@routes/indexPlane';
-import { storageRouter } from '@routes/storage';
-import { userRouter } from '@routes/user';
-import { utilitiesRouter } from '@routes/utilities';
+import { customStorageRouter, CustomStorageRouter } from '@routes/customStorage';
+import { indexRouter, IndexRouter } from '@routes/indexPlane';
+import { storageRouter, StorageRouter } from '@routes/storage';
+import { userRouter, UserRouter } from '@routes/user';
+import { utilitiesRouter, UtilitiesRouter } from '@routes/utilities';
 
 export { createContext } from '@routes';
 
-export const appRouter = t.router({
+export const appRouter = t.router<{
+    customStorage: CustomStorageRouter;
+    index: IndexRouter;
+    storage: StorageRouter;
+    user: UserRouter;
+    utilities: UtilitiesRouter;
+}>({
     customStorage: customStorageRouter,
     index: indexRouter,
     storage: storageRouter,