+++ examples/snap-chapi-example/CHANGELOG.md
@@ -1,5 +1,75 @@
 # @learncard/snap-chapi-example
 
+## 1.0.233
+
+### Patch Changes
+
+-   Updated dependencies [[`1ed5313935264890917c6ddf19249ada91d1e524`](https://github.com/learningeconomy/LearnCard/commit/1ed5313935264890917c6ddf19249ada91d1e524)]:
+    -   @learncard/core@9.3.29
+    -   @learncard/meta-mask-snap@1.0.80
+    -   @learncard/helpers@1.1.17
+    -   @learncard/react@2.7.89
+
+## 1.0.232
+
+### Patch Changes
+
+-   Updated dependencies [[`0b10e74bac01e5232031539a8f88b9b63aee847b`](https://github.com/learningeconomy/LearnCard/commit/0b10e74bac01e5232031539a8f88b9b63aee847b)]:
+    -   @learncard/react@2.7.88
+
+## 1.0.231
+
+### Patch Changes
+
+-   Updated dependencies [[`aaa05036a29e1071bf1b93e9bfdecdc89fb3d7c3`](https://github.com/learningeconomy/LearnCard/commit/aaa05036a29e1071bf1b93e9bfdecdc89fb3d7c3)]:
+    -   @learncard/core@9.3.28
+    -   @learncard/meta-mask-snap@1.0.79
+    -   @learncard/react@2.7.87
+
+## 1.0.230
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/react@2.7.86
+
+## 1.0.229
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.27
+    -   @learncard/helpers@1.1.16
+    -   @learncard/react@2.7.85
+    -   @learncard/meta-mask-snap@1.0.78
+
+## 1.0.228
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.26
+    -   @learncard/helpers@1.1.15
+    -   @learncard/react@2.7.84
+    -   @learncard/meta-mask-snap@1.0.77
+
+## 1.0.227
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.25
+    -   @learncard/helpers@1.1.14
+    -   @learncard/react@2.7.83
+    -   @learncard/meta-mask-snap@1.0.76
+
+## 1.0.226
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/react@2.7.82
+
 ## 1.0.225
 
 ### Patch Changes