+++ services/learn-card-network/simple-signing-service/CHANGELOG.md
@@ -1,5 +1,86 @@
 # @welibraryos/lca-api-service
 
+## 1.0.13
+
+### Patch Changes
+
+-   [#682](https://github.com/learningeconomy/LearnCard/pull/682) [`1ed5313935264890917c6ddf19249ada91d1e524`](https://github.com/learningeconomy/LearnCard/commit/1ed5313935264890917c6ddf19249ada91d1e524) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Bump tsc-alias
+
+-   Updated dependencies [[`1ed5313935264890917c6ddf19249ada91d1e524`](https://github.com/learningeconomy/LearnCard/commit/1ed5313935264890917c6ddf19249ada91d1e524), [`1ed5313935264890917c6ddf19249ada91d1e524`](https://github.com/learningeconomy/LearnCard/commit/1ed5313935264890917c6ddf19249ada91d1e524)]:
+    -   @learncard/core@9.3.29
+    -   @learncard/types@5.7.0
+    -   @learncard/did-web-plugin@1.0.66
+    -   @learncard/didkey-plugin@1.0.40
+    -   @learncard/didkit-plugin@1.5.16
+    -   @learncard/expiration-plugin@1.1.45
+    -   @learncard/learn-card-plugin@1.1.44
+    -   @learncard/vc-plugin@1.1.45
+
+## 1.0.12
+
+### Patch Changes
+
+-   Updated dependencies [[`aaa05036a29e1071bf1b93e9bfdecdc89fb3d7c3`](https://github.com/learningeconomy/LearnCard/commit/aaa05036a29e1071bf1b93e9bfdecdc89fb3d7c3)]:
+    -   @learncard/core@9.3.28
+    -   @learncard/did-web-plugin@1.0.65
+    -   @learncard/didkey-plugin@1.0.39
+    -   @learncard/didkit-plugin@1.5.15
+    -   @learncard/expiration-plugin@1.1.44
+    -   @learncard/learn-card-plugin@1.1.43
+    -   @learncard/vc-plugin@1.1.44
+
+## 1.0.11
+
+### Patch Changes
+
+-   Updated dependencies [[`8957eda698be3222588d0b56834d7ef5622057fb`](https://github.com/learningeconomy/LearnCard/commit/8957eda698be3222588d0b56834d7ef5622057fb)]:
+    -   @learncard/didkit-plugin@1.5.14
+    -   @learncard/learn-card-plugin@1.1.42
+    -   @learncard/vc-plugin@1.1.43
+    -   @learncard/expiration-plugin@1.1.43
+
+## 1.0.10
+
+### Patch Changes
+
+-   Updated dependencies [[`b04788beded98db2fb3827c94e0810943b7f698a`](https://github.com/learningeconomy/LearnCard/commit/b04788beded98db2fb3827c94e0810943b7f698a), [`c9633be1bb84eb620502543383eb17b34be1586e`](https://github.com/learningeconomy/LearnCard/commit/c9633be1bb84eb620502543383eb17b34be1586e)]:
+    -   @learncard/types@5.6.14
+    -   @learncard/didkit-plugin@1.5.13
+    -   @learncard/core@9.3.27
+    -   @learncard/did-web-plugin@1.0.64
+    -   @learncard/didkey-plugin@1.0.38
+    -   @learncard/learn-card-plugin@1.1.41
+    -   @learncard/vc-plugin@1.1.42
+    -   @learncard/expiration-plugin@1.1.42
+
+## 1.0.9
+
+### Patch Changes
+
+-   Updated dependencies [[`319bd3a589e3529d162825d8f6b97268c44060f4`](https://github.com/learningeconomy/LearnCard/commit/319bd3a589e3529d162825d8f6b97268c44060f4)]:
+    -   @learncard/types@5.6.13
+    -   @learncard/core@9.3.26
+    -   @learncard/did-web-plugin@1.0.63
+    -   @learncard/didkey-plugin@1.0.37
+    -   @learncard/didkit-plugin@1.5.12
+    -   @learncard/learn-card-plugin@1.1.40
+    -   @learncard/vc-plugin@1.1.41
+    -   @learncard/expiration-plugin@1.1.41
+
+## 1.0.8
+
+### Patch Changes
+
+-   Updated dependencies [[`ef22869f9e9b99200e46cbd82dfc4c41558afcd1`](https://github.com/learningeconomy/LearnCard/commit/ef22869f9e9b99200e46cbd82dfc4c41558afcd1)]:
+    -   @learncard/types@5.6.12
+    -   @learncard/core@9.3.25
+    -   @learncard/did-web-plugin@1.0.62
+    -   @learncard/didkey-plugin@1.0.36
+    -   @learncard/didkit-plugin@1.5.11
+    -   @learncard/learn-card-plugin@1.1.39
+    -   @learncard/vc-plugin@1.1.40
+    -   @learncard/expiration-plugin@1.1.40
+
 ## 1.0.7
 
 ### Patch Changes